Vuex: Create separate "utils" file for Vuex storage, update usages
