+++ src/vuex/store/constants.ts
@@ -1,5 +1,1 @@
-import { getModuleKey } from '@/utils';
-
 export const FilmsModuleName: string = 'films';
-
-export const getFilmModuleType = (key: string) => getModuleKey(FilmsModuleName, key);
