+++ src/vuex/store/utils.ts
@@ -1,0 +1,4 @@
+import { FilmsModuleName } from '@/vuex/store/constants';
+import { getModuleKey } from '@/utils';
+
+export const getFilmModuleType = (key: string) => getModuleKey(FilmsModuleName, key);
